add ability to specify provider profile for groups
Carves off the ProviderProfile into its own separate thing and modifies
the deployment-config structure to allow specifying a default provider
profile to use for the deployment as well as overrides for any group of
providers. This way, deployment configurations can be created that
override, for instance, just one rack of compute in a particular site to
have, say, SSD instead of HDD for disk storage.

Issue #2