provider generation increment to claim execution
Adds the step of incrementing the generation of all providers involved
in any allocation item of the claim during execution. This allows us to
retry the claim (and check capacity is not exceeded) when another thread
consumes inventory on a provider involved in our claim in between the
initial read of provider state and the end commit of the transaction.